+++ src/css/SuportCard.css
@@ -1,5 +1,5 @@
 .support_card {
-  width: 200px;
+  width: 220px;
   height: 300px;
   display: flex;
   flex-direction: column;
